refactor(): move document&collection dto & responses to it's package in schemas
