+++ app/schemas/document/document_upload_response.py
@@ -1,7 +1,7 @@
 from pydantic import BaseModel
 
-from app.schemas.collection_dto import CollectionDTO
-from app.schemas.document_response import DocumentResponse
+from app.schemas.collection.collection_dto import CollectionDTO
+from app.schemas.document.document_response import DocumentResponse
 
 
 class DocumentUploadResponse(BaseModel):
